Create an advanced screenrecord system [1/2]
part 2/2 is in frameworks native

Multi ScreenRecord tile, blinking stop button, screenrecord service improvements [1/2]

single tap to toggle quality:
- 480*800 resolution and 1.5Mbps
- default resolution (720p) and 4Mbps
- default resolution (720p) and 8Mbps
long press to start recording

Show also a blinking red button at bottom right of the screen
that can be tapped to stop the recording instead of using
the notification.

Also improve the screenrecord service and be sure
to stop it when we are done.

part 2/2 is in framework/av tree, it just sets max recording
duration to 30 minutes instead of 3

Change-Id: I9a0161e6ddafb721421585c7f16bba4b30319425

Screenrec: let hide the red stop button but keep it clickable

thanks to @KreAch3R for this great idea.

Here we add an option to the screenrecord notification to hide
the red stop hint at bottom right so it doesn't appear in our
recordings, but it's still clickable to stop the service.

Change-Id: I1b93b10f07c415c0149097635d78634e0d227932

Screenrecord: show a Share/Delete notification when stopped

also fix MediaScannerConnection memory leak due to aosp bug:
https://issuetracker.google.com/issues/37046656
manually connecting and disconnecting the client

Change-Id: Ia8ea68dd4c4957d14294d0c1bf24f15daac122cb

Screenrecord red hint: long press on it to move to the opposite side

we can already hide it from notification but we still keep it clickable,
but in some cases we would need to click for example on the keyboard
search/enter button covered by the hint.

Change-Id: Ibd541987a0c3ea4be6ea7016b38f248d83431de1

Better strings for screenrec tile

Change-Id: I88822f736d07e05d2464275e2d07b19a6a86fac7

Screenrecord notification: tap on it to open the video

Change-Id: Ib92c9d37ad9ea107b01bca630ae9231b69ebbf18

Screenrecord: show chrono and total recording duration/size

Change-Id: Ic5c72348cf5848d6252f32ef49cec54a375e9f2f

Simplify successful screenrecord notification strings

Change-Id: I1c1396de040d82b8fe92a08de1207523ef4866eb