base: Introduce lockscreen clock/date styles [1/2]
Inspired by OnePlus and their recent options to switch lockscreen
clock styles on the OP6, I took it upon myself to do the same with our ROM.

This isn't as fancy as OnePlus' because is just using a list preference instead
of adding in all the eye candy but it gets the job done!

Below is a run down of what took place...

- Mirrored the analog clock class to avoid issues with any apps that may
or may not be utilizing. Google deprecated this class back in 2015 but there's
still a chance that something may go wrong some where. Better safe than sorry.

/core/java/android/widget/AnalogClock.java@o8x

- Added the analog clock to the keyguard and made it optional
- Manipulated the clockview format for both 12 and 24 hour to achieved the 'sammy' style
- Made the sammy style optional and created a switch case for all 3.
- Set padding, singeline and other layout changes programmatically to minimize the jank
- Reworked/cleaned up some existing code (Huge thanks to @ezio84)
- Created bold sammy by using html formatting, based off the sammy one
- Created bold digital by using html formatting, based off the digital one
- Vectorized the analog clock images so we can theme
- Modified Analog clock class to fix ambient display issues (@ezio84)

For right now there's five clock styles:

Clock styles---
- Default
- Digital (bold)
- Analog
- Sammy
- Sammy (bold)

Change-Id: I57d38ab979a203c0c5b5e6ace17fcb52c5e8fff0
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

Keyguard items: code improvements

- don't update settings on each time change or keyguard state change
event but only when the user changes settings
- clean up

Change-Id: Id7ecb75c60e924de733fed07915ef15e6512a9d2
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

Remove the 0 from the 12 hour format from the 'Digital (bold)' option

This was an oversight when we first pushed the lockscreen clock/date styles.
The 12 clock format should never have a 0 in front of it. That type of format
is reserved for the 24 clock format.

Example of the fail:

If it was 9pm, the clock would show 0900

Change-Id: I8d1b8f785826bfe69f26c6a909421e0911ed6283
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

base: Ensure the clockview is aligned correctly each time

Previously if you switched from Analog Clock to any of the Sammy style
clock views, you would notice sometimes that the view wasn't aligned
correctly. Shutting the device off and turning it back on would fix this
issue but that's just not the right fix for this.

- Add back letterSpacing in clock view. It was removed by mistake

Change-Id: Iae3d8fd875bef11f485c8ec8167861bb63fdef55
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

Lockscreen/ambient analog clock: improve design

add hours indicators and set hours hand to grey for
ambient display

Change-Id: I6aed62c222fec73831b25abf20a69f9df864ee64
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

base: Lockscreen clock styles: Move everything to the keyguard [1/2]

- Add DU analog clock
- Use XML attributes to keep redundancy down. Same result, less code :-)

Change-Id: I62f47ead59bc472f0dd5f31a1f16283e434847fc
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

Lockscreen items/clocks: KeyguardSliceView improvements/cleanup

don't check LOCKSCREEN_INFO more times but refresh its value from StatusBar observer

setDarkAmount gets called a lot of times (and consecutively in the same
keyguard session to get different dark float values), so instead of calling
showSlice() (which does a lot of stuff) there we can just set the needed
visibility for the slice row.

fix mRow.setVisibility call in showSlice (it wasn't handling the case where
no items are available, in that case as per original aosp code the row must
be set to GONE).

Change-Id: I76d7e9010d5681a4d0e7bc1dc20475844c5e154d
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

Lockscreen clock styles: Fix bottom margins & ambient display/AOD visibility

With the commit below, we set the clock using XML attributes vs having to clone
the class and all that jazz. The problem was that we forgot to the same to the
amibent display drawables. This fixes that.

Also the bottom margins for the analog clocks were a bit too small so we've
double that (using keyguard_security_view_top_margin dimens)

Change-Id: I65a401180a64836b71b8d330028e6f4d449834c5
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

Always use standard small clock when pulsing on new notifications

    this is a quick fix for custom clock glitches when the device pulses
    on a new notification and at the same time it keeps what Google wanted
    in this case: show a small clock and give more importance to the
    notification itself

    TODO: handle custom clocks in onLayoutChange shouldAnimate part

Change-Id: I3f9aa8849ac98d498c7d31626458a8b1f4bf7733
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

base: Lockscreen analog clock: improve minute hands looking

it was uncentered

https://i.ibb.co/Ny0pVj5/Screenshot-20190106-160029.png

Change-Id: I8766dea1ffeff44581326f60645b08c5a6e87d6f

lockscreen : differentiate hour from minutes in digital bold style clock

Change-Id: I88a1244972ad0d0e28eac4fbe9d543861b965e61
Signed-off-by: Shreyansh Lodha <slodha96@gmail.com>
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>

Add sammy accent clock and a variant

Inspired on a KLWP design i saw someday, this just adds accent on time because it looks nice
The second design is an attempt to make it kind of similar to some designs i saw for Rainmeter

Co-authored-by: ezio84 <brabus84@gmail.com>
Co-authored-by: ElDainosor <eldainosor@gmail.com>
Co-authored-by: Shreyansh Lodha <slodha96@gmail.com>
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>
Change-Id: Iddae85abfc1d7c0518d8ef61731c592a4809f2e2