HeadsUpAppearanceController: allow heads-up to draw beyond bounds
On some devices with huge `rounded_corner_content_padding`, the icon of
heads-up notifications will be clipped or even completely removed from
view. This is caused by `HeadsUpStatusBarView` trying to align the
heads-up icon with where the icon would be on a normal notification
card, which would end up outside of the status bar contents view when
`rounded_corner_content_padding` is set to beyond a certain value.

This patch fixes the problem by allowing heads-up content to overflow
the normal status bar content region when it is shown. As long as the
rounded corner is not too huge, the aligning position of the icon should
not be a problem of its own.

The patch has already been sumitted to AOSP and cherry-picked into
Android Q. See
https://android-review.googlesource.com/c/platform/frameworks/base/+/967972
for a conversation at AOSP (the patch set isn't marked as merged because
it was cherry-picked manually into the otherwise frozen Android Q branch)

Change-Id: I813f8ffc3c3bf19a7b83881f60ac5a5dd912afc0